ComponentFilesUtils.js is now self-contained
